Avoid reusing timeline for interrupted promotion
Each Postgres instance (that GPDB segments are) should use own timeline
to exclude WAL segment names interferension if derived from the common
ancestor. When we derive the new instance from backup or promote a
replica, the Postgres searches the next free timeline number to identify
the given reincarnation. Postgres perists information about selected
timeline in the Checkpoint record or associates it with End of Recovery
record in the controldata. Also, its remembers switch point to the
history file. These two entities helps the Postgres to handle timeline
switch in case of failure and consequent recovery. But Postgres used to
remove signal files earlier than persisting new timeline info. As a
result, the Postgres had no way to remember about promotion or to try
again if something goes wrong soon. So, Postgres carried out the common
crash recovery and continued the previous timeline. As a result, the
instance may overwrite next WAL segments and push it to the archive with
the same name with unpredictable results depending on archive_command
implementation. But even without archiving, possible presence of this
orphaned timeline in the history file may lead to unexpected error
during the next replica switching because LSN of such switch point is
before latest checkpoint record created on the reused timeline.

The optimal solution is to retry interrupted promotion. FTS is ready to
do it, at least. To achieve this goal and fall into archive recovery
again we should delay removing signal files until we persist the new
timeline information. The only insignificant negative outcome for this
approach is possible assignment of the one more timeline identity, if
we fall right after history file creation.