Partially revert "Move Orca's GUCs into the extension (#1167)" (#1191)
This partially reverts commit 341d342a1daedb928a5b3ca081ba00c90aba94ce: it 
returns back all GUCs into the core (except 'optimizer_partition_selection_log', 
which is not used anywhere), but preserves GUC infrastructure inside the 
extension and preserves changes in minirepro utility tool. 

Reason: moving the GUCs into the extension has broken the ability to set them 
in 'postgresql.conf' file. It is so, because the config file is parsed first 
time before the shared libs are loaded. Thus, if any of the moved GUCs was set 
in the config file, parsing failed with error "unrecognized configuration 
parameter". 

GPDB can handle GUCs from extensions properly, if the GUCs have a qualified name 
(e.g. "<extension_name>.<guc_name>"). So, in order to move the Orca related GUCs 
into the extension, we need to rename them in a way "optimizer_guc_name" -> 
"orca.optimizer_guc_name". That will break backward compatibility, which we want 
to avoid in current release. Therefore, we leave Orca related GUCs inside the 
core in the current release, and we'll move them along with the name updating in 
the next major release.