Unify the projection of the inner node of CorrelatedNLJoin
Previously, the projection of the internal node was redefined only for
EopPhysicalCorrelatedLeftOuterNLJoin, but not for
EopPhysicalCorrelatedInLeftSemiNLJoin and
EopPhysicalCorrelatedNotInLeftAntiSemiNLJoin. This was due (most likely)
to the fact that only for EopPhysicalCorrelatedLeftOuterNLJoin the
<dxl:TestExpr> handler could correctly process the left node of
<dxl:TestExpr> with one parameter.
Since the TestExpr handler can now support trees in the left node with
multiple parameters, the projection can be unified for the rest of the CorrelatedNLJoins.
The logic for determining the projection of the columns of the internal node remains the
same. Assert with the dxl_subplan_type definition remains the same.