Don't modify prepared statement by CTAS from it (#1215)
Creating a table from an execution of a prepared statement modifies the
parentStmtType attribute of the cached query of that statement, which results
in invalid plans for subsequent executions of that prepared statement. This
patch fixes the issue by marking the query as CTAS while leaving the cache
intact, namely, we now modify the parentStmtType attribute only on the copy of
the query that is subsequently used for planning, but not on the cached query
that remains for the prepared statement.

Ticket: ADBDEV-7019
(cherry picked from commit 8b17bffc43dcfbf44b813f2f18b5c3b87c418f1b)

Changes from original commit:
1) resolved conflicts in tests
2) adapt tests output