Fix creating table by EXPLAIN ANALYZE CTAS from prepared statement (#1197, #1226)
When creating a table from a select statement, GPDB uses the following hack.
The intorel_initplan function is called in InitPlan before ExecutorRun in
ExecCreateTableAs so that the table is created earlier, otherwise the current
MPP framework will fail. This only happens if the queryDesc->plannedstmt
structure has the intoClause field filled in. For EXPLAIN ANALYZE CTAS from a
prepared statement, this field was forgotten to be filled in, the
intorel_initplan function was not called, the table was not created, and the
custom DestReceiver DR_intorel was not initialized correctly. In particular, it
had a zero rel field, which was later unconditionally dereferenced in the
intorel_startup_dummy function. But even if a check was added before this
dereference, the query still did not create the table. This patch solves these
problems by correctly setting the intoClause field of the plannedstmt structure
in the ExplainOnePlan function, before calling the ExecutorStart function,
which later calls InitPlan, which then calls intorel_initplan.

Ticket: ADBDEV-6420, ADBDEV-7099