Fix parsing timeline history file
In case of non-linear order of backups we can get case, when history file
contains information about timelines, which were after timeline, when such
backup was created. It means, that WAL files may be read from the incorrect (for
the backup) timelines, and such backup may no be recovered.

To fix it, during recovery it is more correct to use timeline from backup label,
which is actual for backup (backup was created, when this timeline was actual).