Fix using multiple references to DML CTE
Commit 20537e17098e824f6296366478821ada8e98b67e forbids multiple using DML CTE.
But if DML CTE is used one time in another CTE and such CTE is used more than
one time, the check from the commit above will not work, because the cterefcount
field contains 1 (and it is correct).

Patch adds modifying the cterefcount field. It allows to check multiple using
DML CTE even if it is used one time and CTE (which uses DML CTE) is used more
than one times.