Resolve conflicts in src/backend/nodes/outfuncs.c (#1255)
When merging with Postgres 12, commit 19cd1cf decided (but it was not explained
why) to remove the switch case from the _outConstraint function, which led to
conflicts with commits 2f65ef5 and 1f71861, which only fixed debug output. So
this patch uses the GPDB version, discarding the incoming changes.

Ticket: ADBDEV-7238