Fix conflict in src/test/isolation/expected/eval-plan-qual.out
Commit ff77d86 added a new test to
src/test/isolation/specs/eval-plan-qual.spec, but this test has been disabled
in GPDB for a long time by commit 598f4b0 and therefore has not been tested for
a long time. When I ran it, I got a segfault. So this patch just brings both
files into compliance with Postgres, and further research on this test is
planned to be done in a separate task.