Fix test src/pl/plpython/expected/python3/plpython_error.out
Commit 99d27e5 added a new test to src/pl/plpython/sql/plpython_error.sql,
while earlier commit 598f4b0 added the name of the function that caused the
error to the function PLy_exec_function. But the name of this function is
already in the context of the error, so this patch brings the error output in
line with vanilla Postgres, including in the test that commit a381d2d added.