ADBDEV-7238: Fix src/backend/storage/smgr/md.c (#1339)
Commit 1b4f1c6 added a register_dirty_segment call to mdcreate for heap tables,
but it does not take into account GPDB-specific logic. mdcreate is called not
only for heap tables but also for AO tables, while register_dirty_segment
should only be called for heap tables. So the patch adds an extra condition to
mdcreate to call register_dirty_segment, and also adds an assertion to
register_dirty_segment to ensure that it is only called for heap tables.

We don't need fsync when creating the AO table because until we insert any data
we won't have any segments.

Also, this patch fixes the crash of the test
src/test/isolation2/expected/ao_same_trans_truncate_crash.out.

Ticket: ADBDEV-7238