Fix the adb8 tests
The README.Ubuntu.bash required for container build is updated.
Additional packages and configuration flags are added to Dockerfile.ubuntu.
SSH key verification is disabled, which increases the stability of behave tests.
The inclusion of dependencies during the build in compile_gpdb.bash is disabled
because we don't use it.
In the script for running ORCA unit tests, the xerces build is disabled because
it is not required.
Fix the use of the old python version in mirrors_mgmt_utils.
In explain_format, fix passing the test with ORCA in tests with jit. Changes are
taken from 7e86a6b.
Update arenadata/readme.md, remove commands for Rocky Linux. Add instructions
for running tests of resource groups.