Resolve conflicts in src/backend/commands/matview.c (#1379)
Commit 2699fc0 added temporary diff table creation to refresh_by_match_merge
function, while earlier commits 0f5f663, 81a20f6 and c654c50 already added
GPDB-specific logic before that place. Unlike upstream, GPDB does not use a
composite type column when updating a materialized view, so the patch first
creates a temporary table with all the columns as in the original materialized
view, and then adds a tid column and a GPDB-specific sid column to it.