Fix unit test CdbDispatchPlan_may_be_interrupted
The unit test CdbDispatchPlan_may_be_interrupted uses the SetSessionUserId
function to initialize the global private variable OuterUserId. This variable
is returned by the GetOuterUserId function to the buildGpQueryString function,
which is called from the cdbdisp_dispatchX function, which in turn is called
from the CdbDispatchPlan function, which is used in the test. Commit 4c9d96f
removed the setting of the global private variable OuterUserId from the
SetSessionUserId function, which caused the test to fail. This patch uses the
SetSessionAuthorization function to initialize the global private variable
OuterUserId, which calls the SetOuterUserId function, which sets the specified
variable. At the same time, the SetConfigOption function was also mocked, which
is now also called from the SetOuterUserId function after the specified commit.