Fix test src/test/regress/expected/matview.out (#1438)
Commit add8bc9 added a new test to file src/test/regress/sql/matview.sql,
however, GPDB does not support performing modifying operations on the executor
slice, so adding index drop was moved after creating the materialized view.
Also, for index creation, the materialized view distribution was explicitly
specified, and the update was performed in utility mode. For ORCA, the output
of the new test was added.