Fix test src/pl/plpython/expected/python3/plpython_trigger.out (#1442)
Commit 4488142 added a new test to src/pl/plpython/sql/plpython_trigger.sql,
but GPDB does not allow updating the distribution column in a trigger, so the
new table recursive_trigger_test was forced to be distributed on a different
column. Also, GPDB does not allow modifying operations to be performed on the
executor slice, so the operations were moved to utility mode.