Fix test src/test/isolation/specs/intra-grant-inplace.spec (#1454)
Commit 796d191 added a new test
src/test/isolation/specs/intra-grant-inplace.spec, but these tests in GPDB are
run in utility mode, in which the ORCA planner does not work, while the
optimizer GUC remains enabled and the checkCanOptSelectLockingClause function
returns false after commit 0e11844, which causes the test to hang. We should
check the optimizer GUC only in dispatch mode.