Fix cost calculation for Left Outer Join in Orca (#1560)
Problem description:
In some cases of particular statistic values (when column statistics are
missing, while relation statistics have non-zero rows count) Orca created not
optimal plan for a query with a join. The not optimal plan had a broadcast
motion, while involved tables were distributed on the join key (so the motion
was excessive).

Root cause:
Cost for CPhysicalLeftOuterHashJoin operator was calculated wrongly in the plan
alternative without a broadcast motion. Thus, the plan alternative with the
motion had less cost, and it was selected.

Reason for wrong cost calculation - in CCostContext() the optimizer tries to
scale down statistics row estimates. It is done by DRowsPerHost(). In case the
child of CPhysicalLeftOuterHashJoin is the motion operator, statistics are not
available, and the rows are simply assumed to be uniformly distributed across
segments. So, M rows are scaled down across N segments to M/N rows per segment.
But in case the child of CPhysicalLeftOuterHashJoin is a table scan, statistics
are available for DRowsPerHost(), so it tries to perform more sophisticated
calculation of rows per segment. It tries to estimate number of distinct values
(NDVs) of distribution columns - if it is smaller than number of hosts, then the
data is assumed to be distributed across a subset of hosts, and rows per segment
are calculated as M/NDVs.

But for the particular statistics values in our problematic case, the value of
NDVs was calculated wrongly and was equal to 1. So, for plan alternative without
a motion, the estimated row number per segment was M/1 rows. Thus, the optimizer
chose the plan with a motion.

Details of NDVs wrong calculation:
The data is taken from a histogram (by CHistogram::GetNumDistinct()). The join
histogram is derived from the histograms of its children. In our case, as column
statistics are not presented, both children's histograms are empty (meaning they
have 0-size buckets array, 0 count of tuples not in the buckets and 0 value of
null frequency - refer to 'CHistogram::IsEmpty()'). But during the derivation of
join histogram, 'CLeftOuterJoinStatsProcessor::AddHistogramsLOJInner()' is
called, which combines null_histogram with the inner_join_histogram. The
null_histogram contains only 1.0 null_frequency for number of rows of the LASJ
on the outer side. And it is not empty histogram, so we combine empty histogram
with non-empty histogram, getting non-empty join histogram, which doesn't
contain statistics for real values and has 1.0 null_frequency. Thus, its
CHistogram::GetNumDistinct() considers there is only 1 distinct value (which is
null) and returns 1.0.

Fix:
1. Improve the check in 'CStatisticsUtils::AddNdvForAllGrpCols()' for the case
when we need to apply default value for the number of distinct values.
Previously it was done only if the histogram was empty. Now we also check for
the result of 'IsColStatsMissing()', so if there are no real column statistics,
we use the default value of the NVDs. Also, special handling is done for bool
columns, as the default NDV value (1000) is not very suitable - we can have at
most 3 values there.
2. null_histogram in CLeftOuterJoinStatsProcessor::AddHistogramsLOJInner() is
now marked as the one with missing column statistics (which is true, as it has
only null frequency).
3. Add handling for m_is_col_stats_missing flag into
'CHistogram::MakeUnionAllHistogramNormalize()'.
4. Add handling for m_is_col_stats_missing flag into
'CHistogram::CopyHistogram()'.
5. Add new test case into CStatsTest test.
6. Update some old test cases, where cost calculation is affected.

Note: 'CStatisticsUtils::MaxNdv()' may require the same update as
'CStatisticsUtils::AddNdvForAllGrpCols()', but it is not used anywhere in the
code right now. So it is left untouched.

(cherry picked from commit 84beaf4fd19ae4c30588c3c87342d10b16a1a5c0)

Changes comparing to the original commit:
MissingColStatsWithLOJ minidump is updated for GPDB 7.