Fix calculation of block number for page verification in gp_replica_check
If the table consists of several segment files, this could lead to an error
when running the gp_replica_check utility. The root cause is that the segment
file number was not taken into account when calculating the checksum of the
page in gp_replica_check. This caused a mismatch between the checksum in the
page (block) header and the calculated checksum of the page. In order to
achieve a checksum match, blockno is passed to PageIsVerified() taking into
account the segment file number.

Additionally, a segmentation fault was eliminated. In get_relfilenode_map(),
the list of segments is explicitly assigned NIL when creating the
RelfilenodeEntry structure. In gp_replica_check.py ".getName()" is replaced
with ".name" to get rid of the deprecated warning.