Fix segfault in copy.sql test by align prepunion with upstream (#1519)
Commit 3373c71 added get_eclass_for_sort_expr to extract RelOptInfo from
root->simple_rel_array by index from newec->ec_relids bitmap, but this bitmap
may contain a special OUTER_VAR with a large value of 65001 for which there are
no elements in root->simple_rel_array. This special OUTER_VAR with a large
value of 65001 was added by commits 598f4b0 and 38d8815. The first commit
598f4b0, when merging with PostgreSQL 9.2beta2, added an OUTER_VAR argument to
generate_append_tlist when calling makeVar, while the original commit 9446718
used 0 in that place. The second commit 38d8815 when merging with PostgreSQL
9.6 (up to a point between 9.6beta4 and 9.6rc1) added an OUTER_VAR argument to
the recurse_set_operations function when calling the generate_setop_tlist
function, while the original commit 3fc6e2d used 0 in that place. So this patch
brings the code upstream and also fixes the assert in the
set_plan_references_input_asserts function to match expectations.