Improve tuple count estimation by 'cdb_estimate_rel_size()' (#1561)
Problem description:
The original problematic scenario from the user's perspective was following:
 - set 'gp_autostats_mode' to 'none';
 - set 'gp_enable_relsize_collection' to 'on';
 - create a table and perform explain for some simple query over the table.
The explain output showed that the row count estimation was several times
different from the actual number. The magnitude of the error depended on the
table storage type, for compressed AO tables it could be >100 times the
difference.

Root cause:
In case statistics are not available, 'cdb_estimate_rel_size()' is used to
estimate the row count for a relation. This function had a bunch of issues:
1. In case 'gp_enable_relsize_collection' is 'on', 'cdb_estimate_rel_size()'
calls 'cdbRelMaxSegSize()' to estimate relation size (and compute tuple count
from it). 'cdbRelMaxSegSize()' returns info only for 1 segment, which wasn't
taken into account. So, due to this issue the resulting number of tuples was
divided by the number of segments. Also, it didn't tale into account table data
skew.
2. 'cdb_estimate_rel_size()' uses under the hood 'pg_relation_size' to retrieve
relation size, and it doesn't take into account if the relation was compressed.
So, due to this issue the resulting number of tuples for AO compressed tables
was divided by the compression ratio.
3. 'cdb_estimate_rel_size()' uses a tuple width value to calculate the row count
estimation from the relation size. Before this patch, tuple width always
considered header's size for heap tuple. But AO tables have different internal
structure, so row estimation for AO tables was affected by the wrong tuple
width.

Fix:
1. Add the new functions 'GetAOCSSegFilesTotalsCluster()' and
'GetSegFilesTotalsCluster()' to retrieve information about AO tables (size and
tuple count).
2. For heap tables use direct call to 'pg_relation_size()' instead of
'cdbRelMaxSegSize()'.
3. Adjust tuple width calculation for AO row-oriented and AO column-oriented
tables in 'cdb_estimate_rel_size()'. We still may need it if we have 'relpages'
values from 'pg_class'.
4. Update expected results file for one explain output in 'gporca' test. Reason:
in this test case an empty partitioned table is created. Some partitions are
AO tables. As 'gp_enable_relsize_collection' is 'off' in this case, though the
table is empty, 'cdb_estimate_rel_size()' uses DEFAULT_INTERNAL_TABLE_PAGES and
estimated row count for each partition is not zero. Before this patch, the row
count estimation for each partition was the same (computed from
DEFAULT_INTERNAL_TABLE_PAGES). But after this patch, row count for AO partitions
has been increased (which is expected as the tuple width for the AO table is
less than for a heap table). And Postgres planner decided to use index scan
instead of sequential scan for the AO partitions, as it provides less cost.

(cherry picked from commit 1f4d9b473693d91c4219a239456aa60ead0f5f7c)

Changes comparing to the original commit:
1. changes in 'gporca' test are no more needed (the affected test was removed);
2. function 'GetAppendOnlyEntryAuxOids()' now has different arguments comparing
to 6x, so calls to this function are updated;
3. plans in test's expected results are updated due to changes in 7x explain
output;
4. test cases were moved into 'qp_misc' test from 'qp_misc_jiras', as the later
now has matchsubs that truncate all cost output of explain completely (refer to
commit c2dbe683e59341385860a48ce5347177dab62f1a for details)