Integrate storage_pending_deletes module into the core code (#1520)
Extend the PendingRelDelete structure. Now it stores the shmemPtr pointer to
the corresponding shared pending deletes list node. The pointer is filled in
RelationCreateStorage as a return value of PdlShmemAdd. The pointer is set as
invalid in RelationDropStorage, because storage dropping can't lead to orphaned
relfilenode.
Replace pfree for the pendingDeletes list entry with a new PendingRelDeleteFree
function which calls PdlShmemRemove before pfree when shmemPtr is valid.
Add initialization of shared memory for the storage_pending_deletes module.
Increase number of LWLocks by MaxBackends to add LWLocks for the module.
Fix Assert in the dsm_create function, because the module is used in the
stand-alone mode, for example, when initdb runs 'postgres --single'.

No special tests are presented in this patch, as the added functionality will
be tested later together with other parts for the orphaned file removal feature.
At this point, it is enough to pass the current standard test set.

Ticket: ADBDEV-7410