The XLOG_SMGR_CREATE_PDL WAL record should be always linked to XID
When a table was created right after transaction beginning, then the
XLOG_SMGR_CREATE_PDL record was added with InvalidTransactionId, because
XLogInsert calls GetCurrentTransactionIdIfAny and the transaction id has not
been got at this moment.

Get the transaction id before the log_smgrcreate function is called.

No special tests are presented in this patch, as the added functionality will
be tested later together with other parts for the orphaned file removal feature.
At this point, it is enough to pass the current standard test set.

Ticket: ADBDEV-7458