Fix lost executor error messages after query cancellation (#1946)
This patch addresses an issue where error messages from the executor are lost if
a query fails (with an error) after a cancellation has been initiated.

PostgreSQL uses two flags to manage query cancellation:

- QueryCancelPending – indicates that the query should be cancelled as soon as
  possible.
- QueryCancelCleanup – set when the query was actually cancelled and cleanup is
  expected. If QueryCancelCleanup is set, error messages are suppressed.

In certain cases, such as in the ADCC, errors from failed queries were not
reported because QueryCancelCleanup was not properly cleared after the
cancellation.

The patch ensures that QueryCancelCleanup is reset appropriately when an error
is detected in the executor, so that legitimate error messages are not
suppressed.

Ticket: ADBDEV-7546

(cherry picked from 3b526c2)

---

The original patch is expanded upon: QueryCancelCleanup is additionaly cleared
in the top-level error handling block as well as in proc_exit_prepare() in
proc_exit().

A new set of tests was made from ground up.

Ticket: ADBDEV-7581

Co-authored-by: Vladimir Sarmin <v.sarmin@arenadata.io>