Fix SyncRepWaitForLSN invocation
The SyncRepWaitForLSN call must occur with holdoff interrupts to prevent the
function from longjmp when the query cancel signal is received, which may lead
to out-of-sync with the mirror. Calling the function in XLogPutNextGxid does not
require additional holdoff interrupts, since before that, LWLock is acquired,
which in turn holdoff interrupts.