Change the need to separate restrictinfo for RTE_VALUES
If the Values ​​Scan of a subquery has parameters, the locus of that path
must change from General to OuterQuery. The locus change occurs if
upperrestrictinfo is not empty (see set_rel_pathlist()). In order to
fill upperrestrictinfo for Values ​​Scan when it has parameters,
rel_need_to_separate_outer_query_restrictinfos() must return "true".
Before this commit, this did not happen and plans with parameterized
Values ​​Scan could fail because Motion could be added above that node.