Fix incorrect usage of validate_and_refill_options
Commit 3534fa2 split out part of default_reloptions in
src/backend/access/common/reloptions.c into a new function build_reloptions,
and changed the memory allocation for the StdRdOptions structure to a
variable-size memory allocation, so that this new function can be used with
different structures, while the GPDB-specific validate_and_refill_options
function, previously called from default_reloptions and now from
build_reloptions, always expects a pointer to an already allocated StdRdOptions
structure as its first argument. Before the mentioned commit, the
default_reloptions function was called only for the APPENDOPTIMIZED, BITMAP,
HEAP and TOAST types, which meant that only for these types the memory for the
StdRdOptions structure was allocated and the validate_and_refill_options
function was called, so this patch adds the corresponding check to the new
build_reloptions function.