Fix incorrect behavior of gp_percentile_disc function with varlena types
GPDB-specific gp_percentile_disc function takes the previous state as a zero
argument, which is NULL on the first iteration, and under some conditions it
returns it, but directly, and not via the PG_RETURN_NULL macro, as a result the
isnull flag can be lost, which leads to dereferencing of the null pointer for
varlena types during further execution. The patch adds the corresponding check
to preserve the isnull flag.