Resolve conflicts in src/backend/storage/lmgr/lock.c
1) Commit c096a80 changed the return type of the LockCheckConflicts function
from int to bool, and removed the STATUS_FOUND define and its usage in
src/backend/storage/lmgr/lock.c, and renamed the local variable status in the
LockAcquireExtended function to found_conflict, while the early commit 61623ce
added a lot of GPDB-specific code in this place.

2) The same commit c096a80 changed the return value of the LockCheckConflicts
function from STATUS_OK to false and from STATUS_FOUND to true, while the early
commits 6b0e52b, 1617960 and 38d8815 already added a lot of GPDB-specific code
in this place.