Add relation_toast_am callback for AO tables
Commit 83322e3 added a new relation_toast_am callback to the TableAmRoutine
structure and its initialization to the heapam_methods global variable, as well
as its use in the table_relation_toast_am function and its use in the
create_toast_table function.

The patch adds initialization of these callbacks to the ao_column_methods and
ao_row_methods global variables of AO tables.