Use shared variable for default number of segments
This patch introduces a mechanism for managing the default number of segments
used in table creation during a rebalance operation in GPDB. By
utilizing a shared variable, the new implementation allows for dynamic
configuration of the segment count based on the ongoing rebalance operation,
while enhancing error handling to prevent issues
when catalog modifications occur.

A new shared variable gp_create_table_rebalance_numsegments is introduced in
gpexpand.h to track the number of segments to use during table creation while
a rebalancing operation is in progress. The shared variable is initialized in
shared memory with appropriate size and get functionality.

Corresponding SQL functions are created in gp_toolkit extension.
The system now checks if a rebalancing operation is active by verifying locks
before allowing modifications to the number of segments. If a lock is not acquired
(indicating that no rebalancing is underway), an appropriate error
message is returned.

Tests from 5b3f506b7bb0be61e8bec6c32e7b16cd764fc652 are updated
to support the new functonality