Resolve conflicts in src/backend/commands/explain.c (#1878)
1) Commit 7559d8e updated the copyrights, while the GPDB-specific copyrights
had already been added to this location.

2) Commit b925a00 added a conditional return to src/backend/commands/explain.c
in the ExplainPrintJIT function, while the earlier commit 32fba77 had already
added a similar return to the same location.

3) Commit 6ef77cf added an add and remove to the ancestors list in
src/backend/commands/explain.c in the ExplainSubPlans function, while the
earlier commit added an UNUSED output to the same location.

4) Commit 6ef77cf renamed the function set_deparse_context_planstate to
set_deparse_context_plan and changed the planstate argument to plan in it, we
need to do the same in GPDB-specific places.