Add relation_toast_am and relation_fetch_toast_slice callbacks for AO tables (#1917)
1) Commit 83322e3 added a new relation_toast_am callback to the TableAmRoutine
structure and its initialization to the heapam_methods global variable, as well
as its use in the table_relation_toast_am function and its use in the
create_toast_table function.

2) Commit ce242ae added a new relation_fetch_toast_slice callback to the
TableAmRoutine structure and its initialization to the global heapam_methods
variable, as well as its use in the table_relation_fetch_toast_slice function
and its use in the toast_fetch_datum and toast_fetch_datum_slice functions.

The patch adds the initialization of these callbacks to the AO table globals
ao_column_methods and ao_row_methods.