Fix segfault after processing OOM
When OOM is received at the cdbcomponent_recycleIdleQE, it will lead to
re-entrance to the cdbcomponent_recycleIdleQE for segdbDesc, which was processed
before OOM. Such case may lead to segfault, because part of resources was clean
earlier, but will be cleaned again.

Also, there are counters: numActiveQEs and numIdleQEs, which are used for
checking for more than zero. numActiveQEs may be less than zero in case of
re-entrance to the cdbcomponent_recycleIdleQE, which may lead to assertion
failed.

This patch modifies cleaning cdb_component_dbs. When AbortTransaction is
processed, cdb_component_dbs should be cleaned without additional allocating
memory and without tries to reusing components.