Reset fault inserted by twophase_pqexecparams
C-language test program twophase_pqexecparams injects fault.
This fault is not resetted within program, so next test script would fail
if it tries to inject same fault before server is restarted.

We add resetting of this fault into calling SQL script, because we want to be sure that fault is resetted regradless of outcome of C program