Fix compilation of setrefs.c, nodeDynamicSeqscan.c, nodeDynamicIndexscan.c, nodeDynamicBitmapHeapscan.c, transform.c and cdbsubselect.c (#1920)
1) Commit 9ce77d7 renamed the varnoold and varoattno fields in the Var
structure to varnosyn and varattnosyn, we need to do the same in GPDB-specific
places.

2) Commit e1551f9 renamed convert_tuples_by_name_map_if_req to
build_attrmap_by_name_if_req and changed return type from AttrNumber to
AttrMap, we need to do the same in GPDB-specific places.

3) Commit 5815696 changed the return type of the addRangeTableEntryForSubquery
function from RangeTblEntry to ParseNamespaceItem, we also need to do this in
the GPDB-specific code.