Resolve conflicts in src/backend/tcop/utility.c (#1911)
Commit 1281a5c in src/backend/tcop/utility.c in the ProcessUtilitySlow function
replaced some code with context filling and calling the AlterTable function,
while an earlier commit c7649f1 added GPDB-specific processing before calling
the transformAlterTableStmt function in the same place. Before the mentioned
commit 1281a5c, the logic of ALTER TABLE was as follows: the ProcessUtilitySlow
function called the transformAlterTableStmt function, which returned a list of
sub-commands for sequential execution (the generateSerialExtraStmts function
unconditionally placed the commands at the end of the list, and the
transformAlterTableStmt function also placed commands at the end of the list in
the case of AT_SetIdentity), for the GPDB, the transformAlterTableStmt function
was called only on the coordinator, and then when each sub-command from the
list was executed on the coordinator, it was dispatched to the segments. After
commit 1281a5c, the logic of ALTER TABLE operation changed significantly: the
transformAlterTableStmt function is now called at the preparation stage, the
list of sub-commands is divided into a list of commands executed before the
main command and a list of commands executed after the main command (the
generateSerialExtraStmts function now places commands in the list before or
after, and the transformAlterTableStmt function in the case of AT_SetIdentity
now places commands only in the list before). For the GPDB, these lists are
filled in on the coordinator and then transferred to the segments, for which a
new fields prepStmts and execStmts are added to the AlterTableCmd structure,
and writing and reading of the prepStmts and execStmts and afterStmts fields
are added to the _outAlterTableCmd, _outAlteredTableInfo, _readAlterTableCmd
and _readAlteredTableInfo functions. Separate prepStmts and execStmts lists for
executing sub-commands on segments are needed because on the coordinator,
sub-commands from the beforeStmts list can be executed at different stages -
preparation or execution. The ProcessUtilityForAlterTable function, when
executed on the coordinator, could dispatch commands to segments, but we do not
need this, because we dispatch all commands at once in the AlterTable function,
so a new global variable gp_dispatch_utility_statement was added.